backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sse-starlette
redis
python-jose[cryptography]
bcrypt
pydantic
python-dotenv
pandas
msal
requests
minio
anthropic
slowapi
pydantic-settings
sqlalchemy[asyncio]>=2.0
alembic
pyotp
greenlet
aiokafka>=0.9.0
fakeredis[aioredis]
pytest-asyncio
httpx
python-multipart
